Doneeeeeeeeee!

+++ src/main/java/org/example/ElectronicTools/CarWash.java
@@ -4,7 +4,7 @@ public class CarWash extends ElectronicTools{
     private int maximumEaterTemperature;
     private int hoseLength;
 
-    public CarWash(String name, String brand, double price, int number, String color, String ID, String companyName, int weight, int powerConsumption, int maximumEaterTemperature, int hoseLength) {
+    public CarWash(String name, String brand, double price, int number, String color, String ID, String companyName, double weight, int powerConsumption, int maximumEaterTemperature, int hoseLength) {
         super(name, brand, price, number, color, ID, companyName, weight, powerConsumption);
         this.maximumEaterTemperature = maximumEaterTemperature;
         this.hoseLength = hoseLength;
